replace OneWire
